# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for the Showcaster AI affiliate video generation SaaS. Handles user authentication, video generation jobs, and image uploads.",
        "title": "Showcaster API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Showcaster Support"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/jobs": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns a paginated list of the authenticated user's jobs ordered by `createdAt` descending.\nOnly jobs belonging to the authenticated user are returned.\n\n**Pagination defaults:** page=1, limit=100\n**Limit range:** 0–100. When limit=0 an empty array is returned alongside the total count.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Jobs"
                ],
                "summary": "List jobs",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "Page number (min 1, default 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "maximum": 100,
                        "minimum": 0,
                        "type": "integer",
                        "description": "Items per page (0–100, default 100)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated job list with metadata",
                        "schema": {
                            "$ref": "#/definitions/dto.ListJobsResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid page or limit value",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Missing or invalid JWT",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "Database unavailable",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/jobs/generate": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Atomically creates one Job record (status=pending) and four Step records\n(Hook, Problem, Solution, Closure — each status=pending) in a single DB transaction,\nthen enqueues the job for async processing by the background worker.\n\n**Accepted enum values:**\n- `productCategory`: beauty, fashion, electronics, health\n- `targetAudience`: man, woman, children, unisex\n- `orientation`: portrait, landscape, square\n- `resolution`: 720p, 1080p, 4k\n\nBoth image URLs must be valid HTTPS URLs (must start with `https://`).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Jobs"
                ],
                "summary": "Submit a video generation job",
                "parameters": [
                    {
                        "description": "Job payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateJobRequest"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Job accepted — poll GET /api/jobs/{jobId} for status",
                        "schema": {
                            "$ref": "#/definitions/dto.CreateJobDocResponse"
                        }
                    },
                    "400": {
                        "description": "Missing or invalid field",
                        "schema": {
                            "$ref": "#/definitions/dto.ValidationErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Missing or invalid JWT",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "Job queue full or database unavailable",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/jobs/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the full status of a job including all 4 pipeline steps and their output video URLs.\n\n**Job status values:** pending → processing → completed | failed\n**Step status values:** pending → processing → completed | failed\n**Progress:** integer 0–100, calculated as (completed steps / 4) × 100\n\n`videoUrl` is `null` for steps that are still pending or processing.\nIt becomes a Cloudinary HTTPS URL once the step is completed.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Jobs"
                ],
                "summary": "Get job status",
                "parameters": [
                    {
                        "type": "string",
                        "example": "550e8400-e29b-41d4-a716-446655440000",
                        "description": "Job ID (UUID v4)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Job details with step statuses",
                        "schema": {
                            "$ref": "#/definitions/dto.GetJobResponse"
                        }
                    },
                    "401": {
                        "description": "Missing or invalid JWT",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Job belongs to a different user",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Job not found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "Database unavailable",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Permanently deletes a job and all its associated steps in a single atomic transaction.\n\n**Constraints:**\n- Only the job owner can delete it.\n- Only jobs with status `completed` or `failed` can be deleted.\n- Attempting to delete a `pending` or `processing` job returns 409.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Jobs"
                ],
                "summary": "Delete a job",
                "parameters": [
                    {
                        "type": "string",
                        "example": "550e8400-e29b-41d4-a716-446655440000",
                        "description": "Job ID (UUID v4)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Job and all steps deleted",
                        "schema": {
                            "$ref": "#/definitions/dto.DeleteJobResponse"
                        }
                    },
                    "401": {
                        "description": "Missing or invalid JWT",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Job belongs to a different user",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Job not found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Job is still pending or processing — cannot delete",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "Database unavailable",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dto.CreateJobDocResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/dto.CreateJobResponseData"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "dto.CreateJobRequest": {
            "type": "object",
            "required": [
                "modelImageUrl",
                "orientation",
                "productCategory",
                "productImageUrl",
                "productName",
                "resolution",
                "targetAudience"
            ],
            "properties": {
                "modelImageUrl": {
                    "type": "string"
                },
                "orientation": {
                    "type": "string",
                    "enum": [
                        "portrait",
                        "landscape",
                        "square"
                    ]
                },
                "productCategory": {
                    "type": "string",
                    "enum": [
                        "beauty",
                        "fashion",
                        "electronics",
                        "health"
                    ]
                },
                "productImageUrl": {
                    "type": "string"
                },
                "productName": {
                    "type": "string",
                    "maxLength": 200
                },
                "resolution": {
                    "type": "string",
                    "enum": [
                        "720p",
                        "1080p",
                        "4k"
                    ]
                },
                "targetAudience": {
                    "type": "string",
                    "enum": [
                        "man",
                        "woman",
                        "children",
                        "unisex"
                    ]
                }
            }
        },
        "dto.CreateJobResponseData": {
            "type": "object",
            "properties": {
                "jobId": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                }
            }
        },
        "dto.DeleteJobResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/dto.DeleteJobResponseData"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "dto.DeleteJobResponseData": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "job deleted successfully"
                }
            }
        },
        "dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "human-readable error message"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "dto.GetJobResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/dto.GetJobResponseData"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "dto.GetJobResponseData": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string",
                    "example": "2024-05-18T10:00:00Z"
                },
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "progress": {
                    "type": "integer",
                    "example": 25
                },
                "status": {
                    "type": "string",
                    "example": "processing"
                },
                "steps": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.StepResponseDoc"
                    }
                }
            }
        },
...
        "dto.JobSummaryDoc": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string",
                    "example": "2024-05-18T10:00:00Z"
                },
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "status": {
                    "type": "string",
                    "example": "completed"
                },
                "thumbnailUrl": {
                    "type": "string",
                    "example": "https://res.cloudinary.com/demo/image/upload/v1716000000/showcaster/images/thumb.jpg"
                }
            }
        },
        "dto.ListJobsResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/dto.ListJobsResponseData"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "dto.ListJobsResponseData": {
            "type": "object",
            "properties": {
                "jobs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.JobSummaryDoc"
                    }
                },
                "limit": {
                    "type": "integer",
                    "example": 10
                },
                "page": {
                    "type": "integer",
                    "example": 1
                },
                "total": {
                    "type": "integer",
                    "example": 42
                }
            }
        },
...
        "dto.StepResponseDoc": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string",
                    "example": "Hook"
                },
                "status": {
                    "type": "string",
                    "example": "completed"
                },
                "videoUrl": {
                    "type": "string",
                    "example": "https://res.cloudinary.com/demo/video/upload/v1716000000/showcaster/videos/hook.mp4"
                }
            }
        },
...
        "dto.ValidationErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "validation failed: Email email"
                },
                "field": {
                    "type": "string",
                    "example": "Email"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter your JWT token as: Bearer \u003ctoken\u003e",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}